+++ trunk/dim3Engine/Sources_View/decals_draw.c
@@ -124,7 +124,7 @@ void decal_render_mark(int stencil_idx,decal_type *decal)
 		cf[2]=decal->tint.b;
 	}
 	else {
-		gl_lights_calc_color((double)decal->x[0],(double)decal->y[0],(double)decal->z[0],cf);
+		gl_lights_calc_color((float)decal->x[0],(float)decal->y[0],(float)decal->z[0],cf);
 		cf[0]*=decal->tint.r;
 		cf[1]*=decal->tint.g;
 		cf[2]*=decal->tint.b;
